Move arm objects to follow subsystem structure
