+++ src/main/java/frc/robot/commands/ArmSetpointCommand.java
@@ -3,7 +3,9 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.constants.RobotConstants.ArmSubsystemConstants.ArmSetpoints;
-import frc.robot.subsystems.ArmSubsystem;
+import frc.robot.subsystems.arm.ArmSubsystem;
+import frc.robot.subsystems.arm.states.ElevatorSetpointState;
+import frc.robot.subsystems.arm.states.RotaryArmSetpointState;
 
 public class ArmSetpointCommand extends SequentialCommandGroup {
     public ArmSetpointCommand(ArmSubsystem arm, ArmSetpoints setpoint) {
